requirements.txt
Django==1.11.7
djangorestframework==3.7.3
lxml==4.1.0
pandas==0.22.0
psycopg2==2.7.3.2
requests==2.18.4